requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
python-dotenv==1.2.2
python-nmap==0.7.1
scapy==2.7.0
sqlparse==0.5.5
tzdata==2026.1
